// msfs-avionics-mirror/src/sdk/utils/geo/GeoInterfaces.ts
/**
 * A set of latitude/longitude coordinates.
 */
export interface LatLonInterface {
  /** The latitude, in degrees. */
  lat: number;

  /** The longitude, in degrees. */
  lon: number;
}

/**
 * A set of latitude/longitude coordinates.
 */
export interface LatLongInterface {
  /** The latitude, in degrees. */
  lat: number;

  /** The longitude, in degrees. */
  long: number;
}